Ignore empty dot source that stopped further graph rendering
Graphviz generates an empty string in this case which d3-graphviz just
ignores without neither issuing any events after 'layoutStart' nor
calling the supplied callback. This causes the application to wait
forever on what it thinks is a pending layout.

It could be argued that this is a bug or lack of functionality in
d3-graphviz, but the now implemented fix seems reasonable anyway.

The problem occurs e.g. when pasting in a completely new DOT source
code in the DOT text editor.