In Chrome, drag the insert panel node shape image as is
The nice WYSIWYG we have in Firefox didn't work since Chrome seems to
drag what's visible on the screen at the position where the specified
image is located, even if the image is covered by another part of the
UI. If the image is off-screen like in the current implementation,
nothing is shown at all. To get something looking like it does on
Firefox we would need to draw the new image somewhere on the screen
where it's visible and drag from there. This might be a solution for
the furture, but not for now.

Fixes https://github.com/magjac/graphviz-visual-editor/issues/1 finally.