Corrected setPersistentState not to evaluate updater fn twice
The updater function was called a second time from the setState
callback, but at that time the state has already been mutated and a
new eveluation may not give the same result.

Instead, the updater function is called once and the result is used to
set the state in local storage and then returned to the setState
function.