Use saved svg as preview when no graph has yet rendered
The svg preview is not generated until it is needed in order not to
waste computational power during normal edit. In the case that the
current graph needs to be saved as a project before it has been
rendered which happens when loading the DOT source from a URL
parameter, the latest saved svg must be used.