Fix select components by drag on Firefox
Fixes https://github.com/magjac/graphviz-visual-editor/issues/102

The mouseup event couldn't stop propagation of the click event to the
div which de-seleced the just selected components. The solution is to
use the click event instead of the mouseup event to finalize the
selection. This event can stop propagation to the div.