Fix zoom stops working when DOT source is cleared
Fixes https://github.com/magjac/graphviz-visual-editor/issues/135.

The zoom functionality stores the original "graph0" <g> element and
applies zoom transforms to it, but that elements was removed when the
DOT source was cleared.

Instead of removing the whole svg, now only the graph elements below
the "graph0" <g> element are removed.