add .js extension to all local imports
An upcoming commit that will set "type": "module" in the package.json
would otherwise give errors like:

ERROR in ./src/pages/index.js 12:0-43
Module not found: Error: Can't resolve '../ButtonAppBar' in '/home/magjac/graphviz-visual-editor/src/pages'
Did you mean 'ButtonAppBar.js'?
BREAKING CHANGE: The request '../ButtonAppBar' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.