fix initial open from browser dialog formatting
When the "Open from browser" dialog was initially opened, the SVG
preview was not sized correctly which led to a malformed dialog. This
was corrected as soon as the user clicked somwhere in the graph table.
The reason for this was that the componentDidUpdate function which is
responsible for the sizing of the SVG is not called initially.

From https://reactjs.org/docs/react-component.html#componentdidmount:

"componentDidUpdate() is invoked immediately after updating
occurs. This method is not called for the initial render."

For some reason, in Material-UI v3.x.x this was not a problem.