add check of inserted node after disabling the node fillcolor attribute
The immediate reason for doing this is to fix some flakiness in this
test suite, that seems to have been caused by an ongoing transition
when this test case ends. Since this is accomplished by waiting until
the transition ends, we might as well check that the node is inserted
in the graph as expected.

This fixes one instance of flakiness that manifested itself as:

  1) Insertion of nodes into the graph
       Default node fillcolor is seleced from the fillcolor picker in the node format drawer:

      Timed out retrying
      + expected - actual

      -'Alice'
      +'n0'