add check of inserted node after disabling the node color attribute
The immediate reason for doing this is to avoid flakiness in this test
suite, that seems to have been caused by an ongoing transition when a
test case ends. Since this is accomplished by waiting until the
transition ends, we might as well check that the node is inserted in
the graph as expected.

No flakiness has yet been observed for this particular test, but it is
assumed to exist, since the behavior is more or less the same as the
tests that were fixed in the previous two commits. This potential
flakiness is believed to have manifested it iself as:

  1) Insertion of nodes into the graph
       Default node color is seleced from the color picker in the node format drawer:

      Timed out retrying
      + expected - actual

      -'Alice'
      +'n0'