remove useless assertion in insert node with multiple styles test
The assertion asserted that an attribute did not exist on an element
that did not not exist. In a future commit that upgrades Cypress to
version 6+ this assertion would fail.

From https://docs.cypress.io/guides/references/changelog#6-0-0:

Cypress now always throws an error when asserting on an element that
doesn't exist in the DOM (unless you're asserting that the element
should not.exist). Assertions such as not.visible, not.contains,
not.have.class, etc will now fail when used on non-existent DOM
elements.