port generate-graphviz-version.js to ESM and temorarily rename with .mjs extension
The .mjs extension is necessary to make Node.js interpret this file as
an ES module. This is a temporary measure until an upcoming commit
sets "type": "module" in the package.json. Without this, we get this
error message:

(node:847131) Warning: To load an ES module, set "type": "module" in the package.json or use the .mjs extension.
(Use `node --trace-warnings ...` to show where the warning was created)
/home/magjac/graphviz-visual-editor/bin/generate-graphviz-version.js:3
import { graphvizVersion } from "@hpcc-js/wasm";
^^^^^^

SyntaxError: Cannot use import statement outside a module