use named imports for@mui/material imports
An upcoming commit that will set "type": "module" in the package.json
would otherwise give errors like:

ERROR in ./src/pages/index.js 8:0-40
Module not found: Error: Can't resolve '@mui/material/Paper' in '/home/magjac/graphviz-visual-editor/src/pages'
Did you mean 'index.js'?
BREAKING CHANGE: The request '@mui/material/Paper' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.