add workaround for @mui/icons-material not published in ESM format
See
https://github.com/mui/material-ui/issues/30671#issuecomment-1819254734.

An upcoming commit that will set "type": "module" in the package.json
would otherwise give errors like:

ERROR in ./src/ButtonAppBar.js 13:0-48
Module not found: Error: Can't resolve '@mui/icons-material/Menu' in '/home/magjac/graphviz-visual-editor/src'
Did you mean 'Menu.js'?
BREAKING CHANGE: The request '@mui/icons-material/Menu' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.