add workaround 1 for react-ace not published in ESM format
See
https://github.com/securingsincity/react-ace/issues/1540#issue-1250613655.

An upcoming commit that will set "type": "module" in the package.json
would otherwise give errors like:

ERROR in ./src/TextEditor.js 9:0-44
Module not found: Error: Can't resolve 'ace-builds/src-noconflict/mode-dot' in '/home/magjac/graphviz-visual-editor/src'
Did you mean 'mode-dot.js'?
BREAKING CHANGE: The request 'ace-builds/src-noconflict/mode-dot' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.