add codecov.yml and specifiy to adjust the base coverage
This avoids getting errors for relative coverage drops that are due to
covered lines being removed.

From
https://docs.codecov.com/docs/commit-status#removed_code_behavior:

adjust_base: consists in adjusting the BASE coverage by
recalculating it, and then comparing it to the unchanged HEAD. The
adjustment is to remove misses, hits and partials from lines that were
removed in the diff from BASE (as if all removals were made already in
BASE) and re-calculate coverage (coverage = hits / (hits + misses +
partials)). This behavior is skipped if target coverage is defined.