force update the text in the text editor only at startup and when editing in the graph
This is a rework of the DOT source handling to avoid a race caused by
constantly feeding back the DOT source to the text editor. This race
seems to have been latent, but an upcoming commit that changes to the
new root API in React 18, which seems much faster, exposes it.

Now, the DOT source is only fed back to the text editor when
specifically requested. This is currenty done at startup and when
editing the graph in the graph view. This is not expected to interfere
with a user editing the text, since it should not be possible to do
that at the same time.

This also removes the now unnecessary workaround for
https://github.com/magjac/graphviz-visual-editor/issues/236
implemented in 0062951f387bfcec7133647f4fcba12e556225ed.