force update the text in the text editor only when explicitly requested
This is a rework of the DOT source handling to avoid a race caused by
constantly feeding back the DOT source to the text editor. Now,
instead the DOT source is only fed back to the text editor when
specifically requested. This is currenty done at startup, when using
the open from browser dialog and when editing the graph in the graph
view. This is not expected to interfere with a user editing the text
in the text editor, since it should not be possible to do that at the
same time.

This also removes workaround for
https://github.com/magjac/graphviz-visual-editor/issues/236
implemented in 0062951f387bfcec7133647f4fcba12e556225ed which is no
longer needed.

This was a latent race bug which was exposed by an upcoming commit
that changes to the new, seemingly faster, root API in React 18.

Without this, the upcoming commit would cause errors like this in the
Cypress end-to-end tests:

  1) Text editor
       A graph is rendered when DOT source code is typed slowly in the text editor:

      Timed out retrying after 10000ms
      + expected - actual

      -'Alce'
      +'Alice'