fix focus removed from the graph when the ctrl key clicked in Firefox
The "Locate Pointer" accessibility feature in Ubuntu highlights the
mouse pointer when the `ctrl` key is pressed. If this feature is
enabled, pressing the control key generates a "blur" event in Firefox,
but not in Chrome. The effect of this was that the graph lost focus
and any subsequently pressed keys were ignored. This effectively
prohibited any keyboard shortcut involving the `ctrl` key to work in
the graph, e.g. copy-and-paste through Ctrl-C and Ctrl-V.

The solution is to keep track of whether the mouse pointer is within
the application or not and ignore blur events and not remove the focus
if it's still withing the application.

This fixes
https://github.com/magjac/graphviz-visual-editor/issues/260.