fix error when drawing undirected edge
A drawn edge was always inserted into the DOT source as a directed
edge ('->') instead of an undirected edge ('--'), even if the graph
was undirected. This resulted in syntax error in line <n> near '->'.

Fixes https://github.com/magjac/graphviz-visual-editor/issues/237.

Note that, while being drawn, the edge is still shown with a normal
arrowhead also in undirected graphs. This is due to
https://github.com/magjac/d3-graphviz/issues/307. However, once
inserted into the DOT source, it will have the correct arrowhead as
defined by the DOT source (explicitly or implicitly).