use correct edgeop for edge being inserted into undirected graph
An edge being inserted into an undirected graph used the edgeop of a
directed edge ('->') instead of an undirected edge ('--'). This
resulted in the edge 'title' text being wrong. This in turn caused the
animated transitioning into the real edge, after the edge was inserted
into the DOT source, to not work.

Related to https://github.com/magjac/graphviz-visual-editor/issues/237.