re-enable resize with fit test in rendering.spec.js
This test was mistakenly not reenabled in the 1.2.0 release (Graphviz
11.0.0) after being temporarily disabled in
80dd84ddfe74e9448c4d429a29817f18f785f81d.

There was a very slight change in `width` when running locally on
Ubuntu 24.04 with an old version of NodeJS, v16.20.2 instead of the
system default v18.19.1. It was not the upgrade from Graphviz 10.0.1
to 11.0.0 as supected in 80dd84ddfe74e9448c4d429a29817f18f785f81d
since they produce identical SVG output. Also Graphviz 12.1.0, which
an upcoming commit will upgrade to, produces the same ouput.